backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pyjwt[crypto]==2.8.0
msal==1.24.1
sqlite-utils==3.35.2
aiosqlite==0.19.0
pydantic==2.5.0
python-dotenv==1.0.0
httpx==0.25.2
shortuuid==1.0.11
validators==0.22.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
